Fixed file structure. Moved css to components' tsx.
