+++ Red River Mart/src/components/common/headerPage.tsx
@@ -1,17 +1,19 @@
-import rrmLogo from '../../assets/rrmLogo.png'
-
+import "./header.css";
 export default function Header() {
-    return (
-        <header className='header'>
-            <ul>
-                <li className="logoList">
-                    <img src={rrmLogo} alt="Red River Mart Logo" className='logo'/>
-                </li>
-                <li>Buy</li>
-                <li>Sell</li>
-                <li>Inbox</li>
-                <li>Account</li>
-            </ul>
-        </header>
-    );
+  return (
+    <header className="header">
+      <div className="logo">
+        <img src="/rrmLogo.png" alt="Red River Mart Logo" />
+      </div>
+
+      <nav className="menu">
+        <ul>
+          <li>Buy</li>
+          <li>Sell</li>
+          <li>Inbox</li>
+          <li>Account</li>
+        </ul>
+      </nav>
+    </header>
+  );
 }
