--- Red River Mart/src/components/footerComponent/Footer.tsx
@@ -1,7 +1,0 @@
-export default function Footer() {
-    return (
-        <footer>
-            <p>&copy; <span>{new Date().getFullYear()}</span> Red River Mart - Created by Alex Carlos, Heuone Castillo & Ranindi Gunasekera. All Rights Reserved.</p>
-        </footer>
-    )
-}
