+++ Red River Mart/src/components/loginComponent/signInComponent.tsx
@@ -1,3 +1,4 @@
+import "./signInComponent.css";
 import rrmLogo from '../../assets/rrmLogo.png'
 
 export default function SignIn() {
